Final Game Product
Game now starts on Menu screen. Has an interactable object. Has collision walls so you can't fall out of the camera view. When Anxiety meter fills it will trigger jumpscare and once the jumpscare audio ends it will go to the game over screen which then goes back to the main menu and the loop can repeat as much as the player wants.